requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
langchain-community
transformers
huggingface-hub
torch  # Required for Hugging Face models
accelerate  # Required for loading large models efficiently

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn